Core functions for parsing format and write
parseToml is simply parsing toml from a file path to a JS object.
formatAndWrite take the data and return a function taking the editor name (check editorFormatters), it then format the data and write to the ouput path.

The export is the core function, parsing the toml then for each editor in the array "[{ vsCode: 'json' }, { atom: 'cson' }]" execute the formatAndWrite function